source/processing-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy==2.1.1
psycopg2-binary==2.9.9
pydantic==2.9.2